Bug fix: Improve last commit

+++ dbmodel/utils/ftrg/gw_trg_cat_feature.sql
@@ -167,24 +167,26 @@ BEGIN
 				VALUES ('||quote_literal(NEW.id)||','||quote_literal(NEW.system_id)||');';
 			END IF;
 			--create child view
-			v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"}, 
-			"data":{"filterFields":{}, "pageInfo":{}, "action":"SINGLE-CREATE" }}';
-			PERFORM gw_fct_admin_manage_child_views(v_query::json);
-				
+			IF NEW.id <>'LINK' THEN
+				v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"},
+				"data":{"filterFields":{}, "pageInfo":{}, "action":"SINGLE-CREATE" }}';
+				PERFORM gw_fct_admin_manage_child_views(v_query::json);
+			END IF;
+
 			--insert definition into config_info_layer_x_type if its not present already
 			IF NEW.child_layer NOT IN (SELECT tableinfo_id from config_info_layer_x_type)
 			and NEW.child_layer IS NOT NULL THEN
 				INSERT INTO config_info_layer_x_type (tableinfo_id,infotype_id,tableinfotype_id)
 				VALUES (NEW.child_layer,1,NEW.child_layer);
 			END IF;
-		
+
 			RETURN NEW;
 
 		ELSIF TG_OP = 'UPDATE' THEN
 			SELECT child_layer INTO v_viewname FROM cat_feature WHERE id = NEW.id;
 			-- update child views
 			--on update and change of cat_feature.id or child layer name
-			IF NEW.id != OLD.id THEN	
+			IF NEW.id != OLD.id THEN
 
 				--if cat_feature has changed, rename the id in the definition of a child view
 				IF NEW.id != OLD.id THEN
@@ -202,16 +204,16 @@ BEGIN
 							v_definition = replace(v_definition,quote_literal(OLD.id),quote_literal(NEW.id));
 
 							--replace the existing view and drop the old trigger
-							EXECUTE 'CREATE OR REPLACE VIEW '||v_schemaname||'.'||NEW.child_layer||' AS '||v_definition||';';   
+							EXECUTE 'CREATE OR REPLACE VIEW '||v_schemaname||'.'||NEW.child_layer||' AS '||v_definition||';';
 							EXECUTE 'DROP TRIGGER IF EXISTS gw_trg_edit_'||lower(NEW.feature_type)||'_'||lower(OLD.id)||' ON '||v_viewname||';';
 
 							EXECUTE 'CREATE TRIGGER gw_trg_edit_'||lower(NEW.feature_type)||'_'||lower(NEW.id)||'
-							INSTEAD OF INSERT OR UPDATE OR DELETE ON '||v_viewname||' FOR EACH ROW EXECUTE 
+							INSTEAD OF INSERT OR UPDATE OR DELETE ON '||v_viewname||' FOR EACH ROW EXECUTE
 							PROCEDURE gw_trg_edit_'||lower(NEW.feature_type)||'('||quote_literal(NEW.id)||');';
 
 						ELSE
-			
-							v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"}, 
+
+							v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"},
 							"data":{"filterFields":{}, "pageInfo":{}, "action":"SINGLE-CREATE" }}';
 							PERFORM gw_fct_admin_manage_child_views(v_query::json);
 
@@ -228,13 +230,13 @@ BEGIN
 
 					END IF;
 
-					
+
 				END IF;
-				
+
 			END IF;
 
 			--if child layer name has changed OR id has changed and p, rename it
-			IF (NEW.child_layer != OLD.child_layer  AND NEW.child_layer IS NOT NULL ) OR 
+			IF (NEW.child_layer != OLD.child_layer  AND NEW.child_layer IS NOT NULL ) OR
 			(NEW.id != OLD.id AND v_isrenameview IS TRUE )THEN
 
 				IF (NEW.child_layer != OLD.child_layer  AND NEW.child_layer IS NOT NULL ) THEN
@@ -248,25 +250,25 @@ BEGIN
 				IF v_viewname IS NOT NULL THEN
 					IF  (SELECT EXISTS (SELECT FROM information_schema.tables WHERE  table_schema = 'SCHEMA_NAME'
 					AND table_name = v_old_child_layer)) = true THEN
-			
+
 						--get the old view definition
 						EXECUTE 'SELECT pg_get_viewdef('''||v_schemaname||'.'||v_old_child_layer||''', true);'
-						INTO v_definition;	
+						INTO v_definition;
 
 						--replace the existing view
-						EXECUTE 'CREATE OR REPLACE VIEW '||v_schemaname||'.'||v_viewname||' AS '||v_definition||';';   
+						EXECUTE 'CREATE OR REPLACE VIEW '||v_schemaname||'.'||v_viewname||' AS '||v_definition||';';
 						EXECUTE 'DROP VIEW IF EXISTS '||v_schemaname||'.'||v_old_child_layer||';';
 						EXECUTE 'DROP TRIGGER IF EXISTS gw_trg_edit_'||lower(NEW.feature_type)||'_'||lower(OLD.id)||' ON '||v_viewname||';';
 
 						EXECUTE 'CREATE TRIGGER gw_trg_edit_'||lower(NEW.feature_type)||'_'||lower(NEW.id)||'
-						INSTEAD OF INSERT OR UPDATE OR DELETE ON '||v_viewname||' FOR EACH ROW EXECUTE 
+						INSTEAD OF INSERT OR UPDATE OR DELETE ON '||v_viewname||' FOR EACH ROW EXECUTE
 						PROCEDURE gw_trg_edit_'||lower(NEW.feature_type)||'('||quote_literal(NEW.id)||');';
 
-						EXECUTE 'UPDATE cat_feature SET child_layer='||quote_literal(v_viewname)||' WHERE id='||quote_literal(NEW.id)||';';	
+						EXECUTE 'UPDATE cat_feature SET child_layer='||quote_literal(v_viewname)||' WHERE id='||quote_literal(NEW.id)||';';
 					ELSE
-		
+
 						EXECUTE 'DROP VIEW IF EXISTS '||v_schemaname||'.'||v_old_child_layer||';';
-						v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"}, 
+						v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"},
 						"data":{"filterFields":{}, "pageInfo":{}, "action":"SINGLE-CREATE" }}';
 						PERFORM gw_fct_admin_manage_child_views(v_query::json);
 
@@ -286,36 +288,36 @@ BEGIN
 
 				--delete definition from config_info_layer_x_type
 				DELETE FROM config_info_layer_x_type where tableinfo_id=OLD.child_layer OR tableinfotype_id=OLD.child_layer;
-				
+
 				--delete definition from sys_table
 				DELETE FROM sys_table where id=OLD.child_layer;
 
 				EXECUTE 'SELECT gw_fct_admin_manage_child_config($${"client":{"device":4, "infoType":1, "lang":"ES"},
 				"form":{}, 	"feature":{"catFeature":"'||NEW.id||'"},
-				"data":{"filterFields":{}, "pageInfo":{}, "view_name":"'||v_viewname||'", 
+				"data":{"filterFields":{}, "pageInfo":{}, "view_name":"'||v_viewname||'",
 				"feature_type":"'||lower(NEW.feature_type)||'" }}$$);';
 
 
 			ELSIF NEW.feature_type !=OLD.feature_type or NEW.system_id !=OLD.system_id THEN
 
 				EXECUTE 'DROP VIEW IF EXISTS '||NEW.child_layer||';';
-				
+
 				--delete configuration from config_form_fields
 				DELETE FROM config_form_fields where formname=NEW.child_layer AND formtype = 'form_feature';
 
 				--delete definition from config_info_layer_x_type
 				DELETE FROM config_info_layer_x_type where tableinfo_id=NEW.child_layer OR tableinfo_id=OLD.child_layer;
-				
-				EXECUTE 'SELECT lower(id) as id,type, concat(''man_'',lower(id)) as man_table, epa_default, 
-				CASE WHEN epa_default IS NOT NULL THEN concat(''inp_'',lower(epa_default)) END AS epa_table 
+
+				EXECUTE 'SELECT lower(id) as id,type, concat(''man_'',lower(id)) as man_table, epa_default,
+				CASE WHEN epa_default IS NOT NULL THEN concat(''inp_'',lower(epa_default)) END AS epa_table
 				FROM sys_feature_cat WHERE id='||quote_literal(NEW.system_id)||';'
 				INTO v_feature;
-				
+
 				v_new_child_layer = replace(NEW.child_layer,lower(OLD.feature_type),lower(NEW.feature_type));
-				
+
 				EXECUTE 'UPDATE cat_feature SET parent_layer = concat(''v_edit_'',lower('||quote_literal(NEW.feature_type)||')),
 				child_layer = '||quote_literal(v_new_child_layer)||' WHERE id = '||quote_literal(NEW.id)||';';
-				
+
 				EXECUTE 'DELETE FROM cat_feature_'||lower(OLD.feature_type)||' WHERE id = '||quote_literal(NEW.id)||';';
 
 				IF lower(NEW.feature_type)='arc' THEN
@@ -336,11 +338,11 @@ BEGIN
 				DELETE FROM config_form_fields where formname=OLD.child_layer AND formtype = 'form_feature';
 
 				--create child view
-				v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"}, 
+				v_query='{"client":{"device":4, "infoType":1, "lang":"ES"}, "form":{}, "feature":{"catFeature":"'||NEW.id||'"},
 				"data":{"filterFields":{}, "pageInfo":{}, "action":"SINGLE-CREATE" }}';
 				PERFORM gw_fct_admin_manage_child_views(v_query::json);
 
-		
+
 			END IF;
 
 			RETURN NEW;
@@ -348,7 +350,9 @@ BEGIN
 		ELSIF TG_OP = 'DELETE' THEN
 
 			-- delete child views
-			EXECUTE 'DROP VIEW IF EXISTS '||OLD.child_layer||';';
+			IF OLD.child_layer IS NOT NULL THEN
+				EXECUTE 'DROP VIEW IF EXISTS '||OLD.child_layer||';';
+			END IF;
 			
 			--delete configuration from config_form_fields
 			DELETE FROM config_form_fields where formname=OLD.child_layer AND formtype = 'form_feature';
